+++ app/models/conversion/voluntary/tasks/sponsored_support_grant.rb
@@ -1,0 +1,7 @@
+class Conversion::Voluntary::Tasks::SponsoredSupportGrant < TaskList::OptionalTask
+  attribute :eligibility
+  attribute :payment_amount
+  attribute :payment_form
+  attribute :send_information
+  attribute :inform_trust
+end